+++ pkg/gui/services_panel.go
@@ -797,10 +797,6 @@ func (gui *Gui) handleServiceExecShell(g *gocui.Gui, v *gocui.View) error {
 	return gui.withServiceInstance(gui.Tr.ExecShell, gui.instanceExecShell)
 }
 
-func (gui *Gui) handleServiceAttach(g *gocui.Gui, v *gocui.View) error {
-	return gui.withServiceInstance(gui.Tr.Attach, gui.instanceAttachConsole)
-}
-
 func (gui *Gui) handleServiceCopyIPv4(g *gocui.Gui, v *gocui.View) error {
 	return gui.withServiceInstance(gui.Tr.CopyIPv4, gui.instanceCopyIPv4)
 }